# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pulso API
  description: >-
    Backend API for Pulso, an operating system for retail kiosks and small businesses.
    Provides session-based multi-tenant authentication, point-of-sale sales recording,
    offline synchronization, and location management.
  version: 0.1.0
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
  /api/auth/register:
    post:
      summary: Register a new business and owner account
      description: >-
        Creates a new User, Tenant, initial Location, TenantMembership (OWNER),
        and establishes an active persistent session via an httpOnly cookie.
      operationId: register
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '201':
          description: Business successfully registered and session established
          headers:
            Set-Cookie:
              schema:
                type: string
                example: pulso_session=...; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '400':
          description: Validation error in submitted fields
        '409':
          description: Email or tenant slug already exists

  /api/auth/login:
    post:
      summary: Log in with credentials
      description: >-
        Validates email and password using Argon2id. On success, establishes an
        opaque session in PostgreSQL and sets the httpOnly pulso_session cookie.
        Returns a generic 401 on invalid credentials to prevent account enumeration.
      operationId: login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Successfully authenticated
          headers:
            Set-Cookie:
              schema:
                type: string
                example: pulso_session=...; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '400':
          description: Malformed request payload
        '401':
          description: Credenciales inválidas

  /api/auth/me:
    get:
      summary: Get current authenticated session details
      description: >-
        Validates the opaque session from the pulso_session cookie and returns
        the active user, tenant, location, and role context.
      operationId: getCurrentUser
      tags:
        - Authentication
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Active session context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          description: Session invalid, expired, revoked, or missing

  /api/auth/logout:
    post:
      summary: Terminate the active session
      description: >-
        Revokes the session in PostgreSQL (revokedAt = NOW()) and clears the
        pulso_session cookie.
      operationId: logout
      tags:
        - Authentication
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Session successfully closed
          headers:
            Set-Cookie:
              schema:
                type: string
                example: pulso_session=; Path=/; Expires=Thu, 01 Jan 1970 00:00:00 GMT
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Sesión cerrada correctamente

  /api/locations:
    get:
      summary: List branches for the active tenant
      description: >-
        Returns all physical locations belonging to the authenticated tenant.
        Tenant ID is strictly derived from the session.
      operationId: listLocations
      tags:
        - Locations
      security:
        - CookieAuth: []
      responses:
        '200':
          description: List of tenant locations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationListResponse'
        '401':
          description: Unauthorized
    post:
      summary: Create a new branch for the active tenant
      description: >-
        Creates a new branch location belonging to the authenticated tenant.
        Restricted to users with OWNER or MANAGER roles. Tenant ID is strictly derived from session.
      operationId: createLocation
      tags:
        - Locations
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationInput'
      responses:
        '201':
          description: Location successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '400':
          description: Invalid payload (e.g. location name too short)
        '401':
          description: Unauthorized or missing session
        '403':
          description: Forbidden (insufficient permissions, requires OWNER or MANAGER role)

  /api/sales:
    post:
      summary: Record a new point-of-sale transaction
      description: >-
        Records an atomic sale transaction. Enforces idempotency via idempotencyKey.
        Derives tenant and location context exclusively from the verified session.
      operationId: recordSale
      tags:
        - Sales
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - idempotencyKey
                - items
                - paymentMethod
                - totalInCents
              properties:
                idempotencyKey:
                  type: string
                  format: uuid
                  example: 550e8400-e29b-41d4-a716-446655440000
                items:
                  type: array
                  items:
                    type: object
                    required:
                      - productId
                      - quantity
                      - unitPriceInCents
                    properties:
                      productId:
                        type: string
                      quantity:
                        type: integer
                        minimum: 1
                      unitPriceInCents:
                        type: integer
                        minimum: 0
                paymentMethod:
                  type: string
                  enum: [CASH, DEBIT, CREDIT, QR]
                totalInCents:
                  type: integer
                  minimum: 0
                amountReceivedInCents:
                  type: integer
                  minimum: 0
      responses:
        '201':
          description: Sale successfully recorded
        '400':
          description: Invalid payload or negative amount
        '401':
          description: Unauthorized

  /api/sync/batch:
    post:
      summary: Synchronize offline sales queue
      description: >-
        Flushes a batch of sales recorded during offline operation.
        Applies idempotency checks per item and scopes operations to session tenant.
      operationId: syncBatch
      tags:
        - Sync
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - batchId
                - operations
              properties:
                batchId:
                  type: string
                  format: uuid
                operations:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Batch processed results
        '401':
          description: Unauthorized

components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: pulso_session

  schemas:
    MembershipRole:
      type: string
      enum: [OWNER, MANAGER, CASHIER]

    UserStatus:
      type: string
      enum: [ACTIVE, DISABLED]

    RegisterInput:
      type: object
      required:
        - businessName
        - locationName
        - ownerName
        - email
        - password
      properties:
        businessName:
          type: string
          minLength: 2
          maxLength: 100
          example: Kiosco El Trébol
        locationName:
          type: string
          minLength: 2
          maxLength: 100
          example: Casa Central
        ownerName:
          type: string
          minLength: 2
          maxLength: 100
          example: Operador Mostrador
        email:
          type: string
          format: email
          maxLength: 255
          example: operador@kiosco.com
        password:
          type: string
          minLength: 12
          maxLength: 128
          example: passwordSegura123!

    LoginInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: operador@kiosco.com
        password:
          type: string
          example: passwordSegura123!

    SafeUser:
      type: object
      required:
        - id
        - email
        - name
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string

    SafeTenant:
      type: object
      required:
        - id
        - name
        - slug
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string

    SafeLocationSummary:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string

    CurrentUserResponse:
      type: object
      required:
        - user
        - tenant
        - location
        - role
        - expiresAt
      properties:
        user:
          $ref: '#/components/schemas/SafeUser'
        tenant:
          $ref: '#/components/schemas/SafeTenant'
        location:
          $ref: '#/components/schemas/SafeLocationSummary'
        role:
          $ref: '#/components/schemas/MembershipRole'
        expiresAt:
          type: string
          format: date-time

    LocationResponse:
      type: object
      required:
        - id
        - tenantId
        - name
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        tenantId:
          type: string
        name:
          type: string
        address:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    LocationListResponse:
      type: array
      items:
        $ref: '#/components/schemas/LocationResponse'

    CreateLocationInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: Sucursal Norte
        address:
          type: string
          maxLength: 255
          nullable: true
          example: Av. San Martín 456